+++ frontend/src/metabase-lib/temporal_unit.ts
@@ -1,0 +1,9 @@
+export const TEMPORAL_UNITS = [
+  "minute",
+  "hour",
+  "day",
+  "week",
+  "quarter",
+  "month",
+  "year",
+] as const;
